test(files): consolidate the two-pass coverage test files
The earlier coverage work left _extra/_extra2 sibling files that were
an artifact of doing it in two passes, not a meaningful split. Fold
each _extra2 file into its topical sibling (same package, helpers are
shared package-wide), with no change to test logic:

  cache_extra2     -> cache_extra        (libsq/files)
  detect_extra2    -> detect_extra        (libsq/files)
  files_extra2     -> download_test       (libsq/files)
  cache_extra2     -> cache_extra         (internal/downloader)
  downloader_extra2-> downloader_extra     (internal/downloader)

Net: five fewer test files. No production code changed.
